Clarifications inspired by live class: .name, TimeoutFilter, FuturePool
Mention that we don't use the value set by .name() in this example, lest
folks paranoidly wonder.

Update TimeoutFilter example to be closer to the real thing.

Don't block your Finagle threads; if you're tempted to, wrap that in a
FuturePool. We don't show a full example, but do point people at the
Finagle README